Factory Method for creating Spring beans
Normally, Spring instantiates a class and performs dependency injection.
However, sometimes it may be necessary to instantiate a class via
another class
In such a scenario, Spring should not create the class on its own but
simply delegate the instantiation to the Factory class.
Spring provides a way to delegate class instantiation to another class
by using factory-method attribute of bean tag